fix(flusher): Retry flush on exit
The flusher keeps a batch of events so retrying doesn't make sense when
we'll flush again with the next event.

But on exit there won't be a next event, so we need to retry there.

Because you can't set the retry behaviour after constructing an Axiom
client, this creates a second one that retries and adds a `RetryOpt` to
the `Flush` method.